adding protoclol

+++ protocol.c
@@ -6,19 +6,16 @@
  */
 
 #include "protocol.h"
-#include "UART.h"
 /* Dependencies */
-#include "task1.h"
-#include "task2.h"
-#include "Clock_adjust.h"
-#include "taskSwitchLed.h"
+/* libs */
+#include "STD_TYPES.h"
 
 void protcole_task(void)
 {
 	u8 u8byte = 0;
 	u8 u8ID, u8Val;
 	/* get one received byte */
-	UART_Receive(&u8byte, 1);
+	UART_Receive(&u8byte, (u16)1);
 	/* Split it to id and value */
 	u8ID = (u8byte >> 4) & 0x0F;
 	u8Val = u8byte & 0x0F;
@@ -33,6 +30,7 @@ void protcole_task(void)
 	break;
 
 	case ID_MOD_SWITCH:
+
 	break;
 
 	case ID_KP_ADJUST:
